backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
groq
beautifulsoup4
trafilatura
googlesearch-python
apscheduler
psycopg2
pytrends
fastembed
